pyproject.toml
[tool.poetry]
name = "project-j"
version = "0.1.0"
description = ""
authors = ["Bharadwaj LS <lsbharadwaj@lrde.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
djangorestframework = "^3.13.1"
Django = "^4.0.1"
drf-social-oauth2 = "^1.1.4"
django-rest-registration = "^0.7.0"
django-cors-headers = "^3.11.0"
django-heroku = "^0.3.1"
gunicorn = "^20.1.0"
django-debug-toolbar = "^3.2.4"
django-phonenumber-field = "^6.1.0"
django-phonenumbers = "^1.0.1"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
django-sslserver = "^0.22"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
